# Optimize Python calls with large amounts of data