--- a/lib/code-caller/code-caller-native.js
+++ b/lib/code-caller/code-caller-native.js
@@ -366,7 +366,10 @@ class CodeCallerNative {
     this._checkState([IN_CALL, EXITING]);
     if (this.state === IN_CALL) {
       this.outputData += data;
-      if (this.outputData.indexOf('\n') >= 0) {
+      // We check for a newline in `data` instead of `outputData` to avoid an
+      // O(n) operation on a potentially very large string. If `data` contains
+      // a newline, then we know that `outputData` contains a newline as well.
+      if (data.indexOf('\n') >= 0) {
         this._callIsFinished();
       }
     }