--- a/lib/code-caller/code-caller-native.js
+++ b/lib/code-caller/code-caller-native.js
@@ -366,9 +366,8 @@ class CodeCallerNative {
     this._checkState([IN_CALL, EXITING]);
     if (this.state === IN_CALL) {
       this.outputData += data;
-      // We check for a newline in `data` instead of `outputData` to avoid an
-      // O(n) operation on a potentially very large string. If `data` contains
-      // a newline, then we know that `outputData` contains a newline as well.
+      // If `data` contains a newline, then `outputData` must contain a newline as well.
+      // We avoid looking in `outputData` because it's a potentially very large string.
       if (data.indexOf('\n') >= 0) {
         this._callIsFinished();
       }